changed size of reb_simulationarchive_blob. This will break backwards compatability.

+++ src/rebound.h
@@ -874,8 +874,8 @@ struct reb_particle reb_get_com_range(struct reb_simulation* r, int first, int l
 // Simulation Archive
 struct reb_simulationarchive_blob {  // Used in the binary file to identify data blobs
     int32_t index;                   // Index of previous blob (binary file is 0, first blob is 1)
-    int16_t offset_prev;             // Offset to beginning of previous blob (size of previous blob).
-    int16_t offset_next;             // Offset to end of following blob (size of following blob).
+    int32_t offset_prev;             // Offset to beginning of previous blob (size of previous blob).
+    int32_t offset_next;             // Offset to end of following blob (size of following blob).
 };
 
 struct reb_simulationarchive{
